+++ app/services/pricing_service.py
@@ -1,10 +1,10 @@
 from sqlalchemy.orm import Session
 from src.app.models.pricing import TradeScheme
-
+from decimal import Decimal
 
 class PricingService:
     @staticmethod
-    def calculate_item_pricing(db: Session, product_id: int, base_price: float, dispatch_qty: int):
+    def calculate_item_pricing(db: Session, product_id: int, base_price: Decimal, dispatch_qty: int):
         """
         Evaluates active trade schemes for a product.
         Returns: (final_price_per_case, free_qty_awarded)
@@ -22,7 +22,7 @@ class PricingService:
         if scheme and dispatch_qty >= scheme.min_qty:
             # 1. Apply Discount
             if scheme.discount_percent > 0:
-                discount_amount = base_price * (float(scheme.discount_percent) / 100)
+                discount_amount = base_price * (Decimal(scheme.discount_percent) / 100)
                 final_price = base_price - discount_amount
 
             # 2. Apply Free Goods (e.g., "Buy 10 get 2 free". If they buy 25, they get 4 free)
